requirements.txt
flask
python-dotenv
splitwise>=3.0.0
Pillow
openai
python-magic==0.4.27
requests==2.31.0
PyPDF2
pillow-heif
python-telegram-bot>=20.0
